Support wgrep in ivy-occur buffers generated from ivy-xref
Ivy has built-in support for
wgrep-mode (https://github.com/mhayashi1120/Emacs-wgrep) in ivy-occur
buffers, assuming wgrep is already installed. However, in order for
wgrep to work the occur buffer needs to be in a specific format. This
adds a function that tells Ivy how to format the ivy-xref occur
buffer, allowing wgrep to work.

However, it is currently subtly broken - although the wgrep edits
are successfully applied, entries in the occur buffer that have been
edited can no longer be clicked on. This seems to be because after
wgrep has edited the line, the underlying (candidate . location) list
that ivy-xref-show-xrefs relies on to jump to the occurence is set to
nil. I haven't figured out how to fix this yet...